third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-invalid-integrity.html

<!DOCTYPE html>
<html>
    <body>
        <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <p>
        This test passes if no 'FAIL' alert appears.
        </p>
        <script src='fail.js' integrity="yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8="></script>
        <script src='fail.js' integrity="ni://#yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8="></script>
        <script src='fail.js' integrity="yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8="></script>
        <script src='fail.js' integrity="ni://sha1;Ywd5te2RvxmuHLzLzZjYo1/0LMw="></script>
    </body>
</html>
